// src/modules/accelerator/translate.enum.ts
export enum AcceleratorKeys {
    NOT_FOUND = "accelerator.NOT_FOUND",
    COMPANY_NOT_FOUND = "accelerator.COMPANY_NOT_FOUND",
    CREATION_ERROR = "accelerator.CREATION_ERROR",
    UPDATE_ERROR = "accelerator.UPDATE_ERROR",
    UPDATED_SUCCESSFULLY = "accelerator.UPDATED_SUCCESSFULLY",
    CANT_DELETE = "accelerator.CANT_DELETE",
    DELETED_SUCCESSFULLY = "accelerator.DELETED_SUCCESSFULLY",
}